Fixed broken redirect for "GET /"
There was a missing "elif" after the code to handle a "GET /". This
caused the server to send two back-to-back responses when handling a
"GET /". The first response was the correct 302 TEMPORARY REDIRECT. The
second response was a 404 NOT FOUND, caused by the missing "elif". This
confuses browsers -- they misinterpret the second response, the 404, as
being in response to whatever their next request is.